video: Moved SoftwareVideoBackend to software decoder crate
