+++ video/software/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod backend;
 pub mod decoder;
